Rename wrappers to aggregators
